chore: More whisper configs, split up whisper/wav2vec2 specific options
