+++ config/config.yaml
@@ -17,11 +17,12 @@ hub_id: alexandrainst/${pipeline_id}
 model_dir: ${dirs.models}/${pipeline_id}
 push_to_hub: false
 
-# Data parameters
-characters_to_keep: 'abcdefghijklmnopqrstuvwxyzæøå0123456789é|'
-
 # Training parameters
 resume_from_checkpoint: false
 ignore_data_skip: false
 wandb: false
 wandb_name: default
+logging_steps: 25
+eval_steps: 1000
+save_steps: 1000
+save_total_limit: 2
