+++ src/coral_models/whisper.py
@@ -192,12 +192,12 @@ class WhisperModelSetup:
             fp16=self.cfg.model.fp16 and not mps_is_available(),
             push_to_hub=self.cfg.push_to_hub,
             evaluation_strategy="steps",
-            eval_steps=self.cfg.model.eval_steps,
-            save_steps=self.cfg.model.save_steps,
-            logging_steps=self.cfg.model.logging_steps,
+            eval_steps=self.cfg.eval_steps,
+            save_steps=self.cfg.save_steps,
+            logging_steps=self.cfg.logging_steps,
             length_column_name="input_length",
             gradient_checkpointing=True,
-            save_total_limit=self.cfg.model.save_total_limit,
+            save_total_limit=self.cfg.save_total_limit,
             load_best_model_at_end=self.cfg.model.early_stopping,
             metric_for_best_model="wer",
             greater_is_better=False,
